contributors/ttwf_bj/samples/CSS Flexible Box/floated-or-absolutely-positioned-inline-flex-computed-display.html

<!DOCTYPE html>
<html>
<head>
	<title>CSS Test: The computed value of 'display' of a floated or absolutely positioned 'inline-flex' element should be 'flex'</title>
	<link rel="author" title="Shen Dayang" href="mailto:archangel.sdy@gmail.com" />
	<link rel="help" href="http://www.w3.org/TR/css3-flexbox/#flex-containers">
	<meta name="flags" content="dom" />
	<meta name="assert" content="The computed value of 'display' of a floated or absoluted positioned 'inline-flex' element should be 'flex'." />
	<script type="text/javascript" src="/resources/testharness.js"></script>
	<script type="text/javascript" src="/resources/testharnessreport.js"></script>
	<style type="text/css">
		#floated {
			display: inline-flex;
			float: left;
		}

		#absolutely-positioned {
			display: inline-flex;
			position: absolute;
		}
	</style>
</head>
<body>
	<div id="log"></div>
	<div id="floated"></div>
	<div id="absolutely-positioned"></div>
	<script type="text/javascript">
		test(function() {
			assert_equals(
				window.getComputedStyle(
					document.getElementById("floated")).getPropertyValue("display"),
				"flex", "The computed value of 'display' of a floated 'inline-flex' element should be 'flex'.");
		}, "floated-inline-flex-computed-display");

		test(function() {
			assert_equals(
				window.getComputedStyle(
					document.getElementById("absolutely-positioned")).getPropertyValue("display"),
				"flex", "The computed value of 'display' of an absolutely positioned 'inline-flex' element should be 'flex'.");
		}, "absolutely-positioned-inline-flex-computed-display");
	</script>
</body>
</html>
